Fix gas service usage in pm

+++ src/interfaces/IPoolManager.sol
@@ -162,7 +162,7 @@ interface IPoolManager {
     function removeVault(uint64 poolId, bytes16 trancheId, address asset) external;
 
     /// @notice TODO
-    function updateCentrifugeGasPrice(uint256 price, uint256 computedAt) external;
+    function updateCentrifugeGasPrice(uint128 price, uint256 computedAt) external;
 
     /// @notice TODO
     function getTrancheToken(uint64 poolId, bytes16 trancheId) external view returns (address);
